requirements.txt
click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
SQLAlchemy-Utils==0.36.1
PyMySQL==0.9.3
requests==2.23.0
pandas==0.25.3
WTForms==2.2.1
Flask-Table==0.5.0
six==1.14.0
urllib3==1.25.9
setuptools==40.8.0
pytz==2019.3
pip==19.0.3
Flask-Babel==1.0.0
psycopg2==2.8.5
pyodbc==4.0.30
urllib3==1.25.9

